+++ torchmetrics/classification/stat_scores.py
@@ -12,7 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 from typing import Any, Callable, Optional, Tuple
-from warnings import warn
 
 import numpy as np
 import torch
@@ -20,6 +19,7 @@ from torch import Tensor, tensor
 
 from torchmetrics.functional.classification.stat_scores import _stat_scores_compute, _stat_scores_update
 from torchmetrics.metric import Metric
+from torchmetrics.utilities import _deprecation_warn_arg_is_multiclass
 from torchmetrics.utilities.enums import AverageMethod, MDMCAverageMethod
 
 
@@ -104,6 +104,9 @@ class StatScores(Metric):
         dist_sync_fn:
             Callback that performs the allgather operation on the metric state. When ``None``, DDP
             will be used to perform the allgather.
+        is_multiclass:
+            .. deprecated:: 0.3
+                Argument will not have any effect and will be removed in v0.4, please use ``multiclass`` intead.
 
     Raises:
         ValueError:
@@ -148,12 +151,7 @@ class StatScores(Metric):
         dist_sync_fn: Callable = None,
         is_multiclass: Optional[bool] = None,  # todo: deprecated, remove in v0.4
     ):
-        if is_multiclass is not None and multiclass is None:
-            warn(
-                "Argument `is_multiclass` was deprecated in v0.3.0 and will be removed in v0.4. Use `multiclass`.",
-                DeprecationWarning
-            )
-            multiclass = is_multiclass
+        multiclass = _deprecation_warn_arg_is_multiclass(is_multiclass, multiclass)
 
         super().__init__(
             compute_on_step=compute_on_step,